# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LanGigaCards API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Statistics/overview": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "languageCode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsOverviewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsOverviewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatisticsOverviewDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Statistics/heatmap": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "languageCode",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeatmapPointDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeatmapPointDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeatmapPointDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HeatmapPointDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "studySeconds": {
            "type": "integer",
            "format": "int32"
          },
          "reviews": {
            "type": "integer",
            "format": "int32"
          },
          "quizAnswers": {
            "type": "integer",
            "format": "int32"
          },
          "xpEarned": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "StatisticsOverviewDto": {
        "type": "object",
        "properties": {
          "period": {
            "$ref": "#/components/schemas/StatisticsPeriodDto"
          },
          "totalStudySeconds": {
            "type": "integer",
            "format": "int32"
          },
          "totalStudyMinutes": {
            "type": "number",
            "format": "double"
          },
          "quizAccuracyPercent": {
            "type": "number",
            "format": "double"
          },
          "quizQuestionsAnswered": {
            "type": "integer",
            "format": "int32"
          },
          "correctAnswers": {
            "type": "integer",
            "format": "int32"
          },
          "reviewCount": {
            "type": "integer",
            "format": "int32"
          },
          "completedLessons": {
            "type": "integer",
            "format": "int32"
          },
          "dueReviews": {
            "type": "integer",
            "format": "int32"
          },
          "currentStreak": {
            "type": "integer",
            "format": "int32"
          },
          "longestStreak": {
            "type": "integer",
            "format": "int32"
          },
          "totalXp": {
            "type": "integer",
            "format": "int32"
          },
          "level": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "StatisticsPeriodDto": {
        "type": "object",
        "properties": {
          "start": {
            "type": "string",
            "format": "date-time"
          },
          "to": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Token değerinizi girin.",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}